initial proof of concept of the 'add to cart' backed by local storage.
cart service is largely copied from some internet examples and it
follows their implementation. It stores the cart as an array of descId
under a specific key to a browser's localstorage. Note that we do not
detect if the browser has the capability. I'll look into that next. If
they have localStorage, we'll use it... If not, perhaps we fall back to
cookies to those users.